+++ src/wrappers.rs
@@ -16,13 +16,17 @@
 //! This crate uses `Boxed*` family as the default, whereas `WithId*`
 //! are type aliases.
 
+use std::fmt;
+use std::marker::PhantomData;
+
 #[cfg(feature = "quickcheck")]
 use quickcheck::{Arbitrary, Gen};
+use serde::de::{Deserialize, Deserializer, Error as DeError, MapAccess, SeqAccess, Visitor};
 
-use error;
+use error::{self, DeErrorKind};
 use identifiable::Identifiable;
 use sized::MtProtoSized;
-use utils::safe_int_cast;
+use utils::{safe_int_cast, safe_uint_cmp};
 
 
 /// A struct that wraps an `Identifiable` type value to serialize and
@@ -30,7 +34,7 @@ use utils::safe_int_cast;
 ///
 /// Note: if you want to attach both id and serialized size to the
 /// underlying data (in this order), see `BoxedWithSize`.
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize, Deserialize)]
+#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
 pub struct Boxed<T> {
     id: u32,
     inner: T,
@@ -64,7 +68,96 @@ impl<T: Identifiable> Boxed<T> {
     }
 }
 
+// Using a custom implementation instead of the derived one because we need to check validity
+// of the deserialized type id __before__ deserializing the value.
+impl<'de, T> Deserialize<'de> for Boxed<T>
+    where T: Deserialize<'de> + Identifiable
+{
+    fn deserialize<D>(deserializer: D) -> Result<Boxed<T>, D::Error>
+        where D: Deserializer<'de>
+    {
+        use identifiable::Identifiable;
+
+        struct BoxedVisitor<T>(PhantomData<T>);
+
+        fn check_type_id<T: Identifiable>(type_id: u32) -> error::Result<()> {
+            let expected_type_ids = T::all_type_ids();
+            if expected_type_ids.iter().find(|&id| *id == type_id).is_none() {
+                bail!(DeErrorKind::InvalidTypeId(type_id, expected_type_ids));
+            }
+
+            Ok(())
+        }
+
+        fn checked_boxed_value<T: Identifiable>(type_id: u32, value: T) -> error::Result<Boxed<T>> {
+            let boxed_value = Boxed::new(value);
+
+            if type_id != boxed_value.id {
+                bail!(DeErrorKind::TypeIdMismatch(type_id, boxed_value.id));
+            }
+
+            Ok(boxed_value)
+        }
+
+        impl<'de, T> Visitor<'de> for BoxedVisitor<T>
+            where T: Deserialize<'de> + Identifiable
+        {
+            type Value = Boxed<T>;
+
+            fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
+                f.write_str("type id and an `Identifiable` value")
+            }
+
+            fn visit_seq<A>(self, mut seq: A) -> Result<Boxed<T>, A::Error>
+                where A: SeqAccess<'de>
+            {
+                let errconv = |kind: DeErrorKind| A::Error::custom(error::Error::from(kind));
+
+                let type_id = seq.next_element()?
+                    .ok_or(errconv(DeErrorKind::NotEnoughElements(0, 2)))?;
+
+                check_type_id::<T>(type_id).map_err(A::Error::custom)?;
+
+                let value = seq.next_element()?
+                    .ok_or(errconv(DeErrorKind::NotEnoughElements(1, 2)))?;
+
+                checked_boxed_value::<T>(type_id, value).map_err(A::Error::custom)
+            }
+
+            fn visit_map<A>(self, mut map: A) -> Result<Boxed<T>, A::Error>
+                where A: MapAccess<'de>
+            {
+                let errconv = |kind: DeErrorKind| A::Error::custom(error::Error::from(kind));
+
+                let type_id = match map.next_key()?
+                    .ok_or(errconv(DeErrorKind::NotEnoughElements(0, 2)))?
+                {
+                    "id" => map.next_value()?,
+                    key => bail!(errconv(DeErrorKind::InvalidMapKey(key.to_owned(), "id"))),
+                };
+
+                check_type_id::<T>(type_id).map_err(A::Error::custom)?;
+
+                let value = match map.next_key()?
+                    .ok_or(errconv(DeErrorKind::NotEnoughElements(1, 2)))?
+                {
+                    "inner" => map.next_value()?,
+                    key => bail!(errconv(DeErrorKind::InvalidMapKey(key.to_owned(), "inner"))),
+                };
+
+                checked_boxed_value::<T>(type_id, value).map_err(A::Error::custom)
+            }
+        }
+
+        deserializer.deserialize_struct("Boxed", &["id", "inner"], BoxedVisitor(PhantomData))
+    }
+}
+
 impl<T: Identifiable> Identifiable for Boxed<T> {
+    fn all_type_ids() -> &'static [u32] {
+        T::all_type_ids()
+    }
+
     fn type_id(&self) -> u32 {
         self.id
     }
@@ -100,7 +193,7 @@ impl<T> Arbitrary for Boxed<T>
 /// A struct that wraps a `MtProtoSized` type value to serialize and
 /// deserialize as a MTProto data type with the size of its serialized
 /// value.
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize, Deserialize)]
+#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
 pub struct WithSize<T> {
     size: u32,
     inner: T,
@@ -130,6 +223,40 @@ impl<T: MtProtoSized> WithSize<T> {
     /// Unwrap the box and return the wrapped value.
     pub fn into_inner(self) -> T {
         self.inner
+    }
+}
+
+// Using a custom implementation instead of the derived one because we need to check validity
+// of the deserialized size against the size hint of a deserialized value.
+impl<'de, T> Deserialize<'de> for WithSize<T>
+    where T: Deserialize<'de> + MtProtoSized
+{
+    fn deserialize<D>(deserializer: D) -> Result<WithSize<T>, D::Error>
+        where D: Deserializer<'de>
+    {
+        let errconv = |kind: DeErrorKind| D::Error::custom(error::Error::from(kind));
+
+        #[derive(Deserialize)]
+        #[serde(rename = "WithSize")]
+        struct WithSizeHelper<T> {
+            size: u32,
+            inner: T,
+        }
+
+        let helper = WithSizeHelper::<T>::deserialize(deserializer)?;
+        let helper_size_hint = helper.inner.size_hint().map_err(D::Error::custom)?;
+
+        if !safe_uint_cmp(helper.size, helper_size_hint) {
+            bail!(errconv(DeErrorKind::SizeMismatch(
+                helper.size,
+                safe_int_cast(helper_size_hint).map_err(D::Error::custom)?,
+            )));
+        }
+
+        Ok(WithSize {
+            size: helper.size,
+            inner: helper.inner,
+        })
     }
 }
 
@@ -165,7 +292,7 @@ impl<T> Arbitrary for WithSize<T>
 /// This struct exists because `Boxed<WithSize<T>>` cannot be created
 /// due to `WithSize<T>` not being `Identifiable` (this restriction is
 /// made on purpose).
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize, Deserialize)]
+#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd, Serialize)]
 pub struct BoxedWithSize<T> {
     id: u32,
     size: u32,
@@ -200,6 +327,119 @@ impl<T: Identifiable + MtProtoSized> BoxedWithSize<T> {
     /// Unwrap the box and return the wrapped value.
     pub fn into_inner(self) -> T {
         self.inner
+    }
+}
+
+// Using a custom implementation instead of the derived one because we need to check validity
+// of the deserialized type id __before__ deserializing the value and also of the deserialized size
+// against the size hint of a deserialized value.
+impl<'de, T> Deserialize<'de> for BoxedWithSize<T>
+    where T: Deserialize<'de> + Identifiable + MtProtoSized
+{
+    fn deserialize<D>(deserializer: D) -> Result<BoxedWithSize<T>, D::Error>
+        where D: Deserializer<'de>
+    {
+        use identifiable::Identifiable;
+        use sized::MtProtoSized;
+
+        struct BoxedWithSizeVisitor<T>(PhantomData<T>);
+
+        fn check_type_id<T: Identifiable>(type_id: u32) -> error::Result<()> {
+            let expected_type_ids = T::all_type_ids();
+            if expected_type_ids.iter().find(|&id| *id == type_id).is_none() {
+                bail!(DeErrorKind::InvalidTypeId(type_id, expected_type_ids));
+            }
+
+            Ok(())
+        }
+
+        fn checked_boxed_with_size_value<T>(type_id: u32,
+                                            size: u32,
+                                            value: T)
+                                           -> error::Result<BoxedWithSize<T>>
+            where T: Identifiable + MtProtoSized
+        {
+            let boxed_with_size_value = BoxedWithSize::new(value)?;
+
+            if type_id != boxed_with_size_value.id {
+                bail!(DeErrorKind::TypeIdMismatch(type_id, boxed_with_size_value.id));
+            }
+
+            let boxed_with_size_size_hint = boxed_with_size_value.size_hint()?;
+            if !safe_uint_cmp(size, boxed_with_size_size_hint) {
+                bail!(DeErrorKind::SizeMismatch(size, safe_int_cast(boxed_with_size_size_hint)?));
+            }
+
+            Ok(boxed_with_size_value)
+        }
+
+        fn next_seq_element<'de, T, A>(seq: &mut A,
+                                       deserialized_count: u32,
+                                       expected_count: u32)
+                                      -> Result<T, A::Error>
+            where T: Deserialize<'de>,
+                  A: SeqAccess<'de>,
+        {
+            let errconv = |kind: DeErrorKind| A::Error::custom(error::Error::from(kind));
+
+            seq.next_element()?
+                .ok_or(errconv(DeErrorKind::NotEnoughElements(deserialized_count, expected_count)))
+        }
+
+        fn next_map_element<'de, T, A>(map: &mut A,
+                                       expected_key: &'static str,
+                                       deserialized_count: u32,
+                                       expected_count: u32)
+                                      -> Result<T, A::Error>
+            where T: Deserialize<'de>,
+                  A: MapAccess<'de>,
+        {
+            let errconv = |kind: DeErrorKind| A::Error::custom(error::Error::from(kind));
+
+            let next_key: String = map.next_key()?
+                .ok_or(errconv(DeErrorKind::NotEnoughElements(deserialized_count, expected_count)))?;
+
+            if &next_key != expected_key {
+                bail!(errconv(DeErrorKind::InvalidMapKey(next_key, expected_key)));
+            }
+
+            map.next_value()
+        }
+
+        impl<'de, T> Visitor<'de> for BoxedWithSizeVisitor<T>
+            where T: Deserialize<'de> + Identifiable + MtProtoSized
+        {
+            type Value = BoxedWithSize<T>;
+
+            fn expecting(&self, f: &mut fmt::Formatter) -> fmt::Result {
+                f.write_str("type id and an `Identifiable` value")
+            }
+
+            fn visit_seq<A>(self, mut seq: A) -> Result<BoxedWithSize<T>, A::Error>
+                where A: SeqAccess<'de>
+            {
+                let type_id = next_seq_element(&mut seq, 0, 3)?;
+                check_type_id::<T>(type_id).map_err(A::Error::custom)?;
+
+                let value = next_seq_element(&mut seq, 1, 3)?;
+                let size = next_seq_element(&mut seq, 2, 3)?;
+                checked_boxed_with_size_value::<T>(type_id, size, value).map_err(A::Error::custom)
+            }
+
+            fn visit_map<A>(self, mut map: A) -> Result<BoxedWithSize<T>, A::Error>
+                where A: MapAccess<'de>
+            {
+                let type_id = next_map_element(&mut map, "id", 0, 3)?;
+                check_type_id::<T>(type_id).map_err(A::Error::custom)?;
+
+                let size = next_map_element(&mut map, "size", 0, 3)?;
+                let value = next_map_element(&mut map, "inner", 0, 3)?;
+                checked_boxed_with_size_value::<T>(type_id, size, value).map_err(A::Error::custom)
+            }
+        }
+
+        deserializer.deserialize_struct(
+            "BoxedWithSize", &["id", "size", "inner"], BoxedWithSizeVisitor(PhantomData))
     }
 }